+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Json serializer control character escape fixed. Some unicode characters
+  were not escaped properly. 
 
 ### Added
 
 - Added optional `categories` member to the `legend-marker`, `legend-label` and `plot-axis-label` events.
+- Remove unused marker selection and selected marker coloring.
+- Remove cursor modification over logo
 
 ## [0.9.3] - 2023-12-20